Moved scorebook related components into a separate folder.
